requirements.txt
langchain
langchain-community
alpha_vantage
python-binance
psycopg2-binary
SQLAlchemy
openai
loguru
pytest
pydantic
pydantic-settings
asyncio
GitPython
docker
fastapi
uvicorn
tenacity
limits
